+++ core/src/com/s31b/castleoffense/game/entity/Entity.java
@@ -20,11 +20,11 @@ abstract public class Entity implements Priceable {
     protected final String description;
     protected final Texture sprite;
     private boolean first = true;
-    protected final float price;
+    protected final int price;
 
     protected Player owner;
 
-    public Entity(EntityType type, String name, String descr, Texture sprite, float price, Player owner) {
+    public Entity(EntityType type, String name, String descr, Texture sprite, int price, Player owner) {
         nextId++;
         this.id = nextId;
         this.type = type;
